pyproject.toml
[project]
name = "dyn-grid-tariffs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "iesopt>=2.6.2",
    "matplotlib>=3.10.1",
    "pandastable>=0.13.1",
    "plotly>=6.0.1",
]
